refactor: inline shared chess primitives into game package
