+++ chessTest/internal/game/abilities/builtins.go
@@ -1,28 +1,26 @@
+// path: chessTest/internal/game/abilities/builtins.go
 package abilities
 
-import (
-	"battle_chess_poc/internal/game"
-	"battle_chess_poc/internal/shared"
-)
+import "battle_chess_poc/internal/game"
 
 func init() {
-	mustRegisterBuiltin(shared.AbilityDoubleKill, game.NewDoubleKillHandler)
-	mustRegisterBuiltin(shared.AbilityScorch, game.NewScorchHandler)
-	mustRegisterBuiltin(shared.AbilityTailwind, game.NewTailwindHandler)
-	mustRegisterBuiltin(shared.AbilityRadiantVision, game.NewRadiantVisionHandler)
-	mustRegisterBuiltin(shared.AbilityUmbralStep, game.NewUmbralStepHandler)
-	mustRegisterBuiltin(shared.AbilitySchrodingersLaugh, game.NewSchrodingersLaughHandler)
-	mustRegisterBuiltin(shared.AbilityGaleLift, game.NewGaleLiftHandler)
-	mustRegisterBuiltin(shared.AbilityQuantumKill, game.NewQuantumKillHandler)
-	mustRegisterBuiltin(shared.AbilityChainKill, game.NewChainKillHandler)
-	mustRegisterBuiltin(shared.AbilityPoisonousMeat, game.NewPoisonousMeatHandler)
-	mustRegisterBuiltin(shared.AbilityOverload, game.NewOverloadHandler)
-	mustRegisterBuiltin(shared.AbilityBastion, game.NewBastionHandler)
-	mustRegisterBuiltin(shared.AbilityTemporalLock, game.NewTemporalLockHandler)
-	mustRegisterBuiltin(shared.AbilityResurrection, game.NewResurrectionHandler)
+	mustRegisterBuiltin(game.AbilityDoubleKill, game.NewDoubleKillHandler)
+	mustRegisterBuiltin(game.AbilityScorch, game.NewScorchHandler)
+	mustRegisterBuiltin(game.AbilityTailwind, game.NewTailwindHandler)
+	mustRegisterBuiltin(game.AbilityRadiantVision, game.NewRadiantVisionHandler)
+	mustRegisterBuiltin(game.AbilityUmbralStep, game.NewUmbralStepHandler)
+	mustRegisterBuiltin(game.AbilitySchrodingersLaugh, game.NewSchrodingersLaughHandler)
+	mustRegisterBuiltin(game.AbilityGaleLift, game.NewGaleLiftHandler)
+	mustRegisterBuiltin(game.AbilityQuantumKill, game.NewQuantumKillHandler)
+	mustRegisterBuiltin(game.AbilityChainKill, game.NewChainKillHandler)
+	mustRegisterBuiltin(game.AbilityPoisonousMeat, game.NewPoisonousMeatHandler)
+	mustRegisterBuiltin(game.AbilityOverload, game.NewOverloadHandler)
+	mustRegisterBuiltin(game.AbilityBastion, game.NewBastionHandler)
+	mustRegisterBuiltin(game.AbilityTemporalLock, game.NewTemporalLockHandler)
+	mustRegisterBuiltin(game.AbilityResurrection, game.NewResurrectionHandler)
 }
 
-func mustRegisterBuiltin(id shared.Ability, ctor func() game.AbilityHandler) {
+func mustRegisterBuiltin(id game.Ability, ctor func() game.AbilityHandler) {
 	if err := Register(id, func() AbilityHandler { return ctor() }); err != nil {
 		panic(err)
 	}
