+++ chessTest/internal/game/abilities/handler_test.go
@@ -1,3 +1,4 @@
+// path: chessTest/internal/game/abilities/handler_test.go
 package abilities
 
 import (
@@ -5,14 +6,13 @@ import (
 	"testing"
 
 	"battle_chess_poc/internal/game"
-	"battle_chess_poc/internal/shared"
 )
 
 func resetRegistry(t *testing.T) {
 	t.Helper()
 	registryMu.Lock()
 	defer registryMu.Unlock()
-	registry = make(map[shared.Ability]HandlerFactory)
+	registry = make(map[game.Ability]HandlerFactory)
 }
 
 func TestRegisterAndNew(t *testing.T) {
@@ -24,11 +24,11 @@ func TestRegisterAndNew(t *testing.T) {
 		},
 	}
 
-	if err := Register(shared.AbilityScorch, func() AbilityHandler { return handler }); err != nil {
+	if err := Register(game.AbilityScorch, func() AbilityHandler { return handler }); err != nil {
 		t.Fatalf("register ability: %v", err)
 	}
 
-	instance, err := New(shared.AbilityScorch)
+	instance, err := New(game.AbilityScorch)
 	if err != nil {
 		t.Fatalf("new handler: %v", err)
 	}
@@ -45,62 +45,62 @@ func TestRegisterDuplicate(t *testing.T) {
 	resetRegistry(t)
 
 	ctor := func() AbilityHandler { return HandlerFuncs{} }
-	if err := Register(shared.AbilityScorch, ctor); err != nil {
+	if err := Register(game.AbilityScorch, ctor); err != nil {
 		t.Fatalf("first registration: %v", err)
 	}
-	if err := Register(shared.AbilityScorch, ctor); !errors.Is(err, ErrDuplicateRegistration) {
+	if err := Register(game.AbilityScorch, ctor); !errors.Is(err, ErrDuplicateRegistration) {
 		t.Fatalf("expected ErrDuplicateRegistration, got %v", err)
 	}
 }
 
 func TestRegisterNilFactory(t *testing.T) {
 	resetRegistry(t)
-	if err := Register(shared.AbilityScorch, nil); !errors.Is(err, ErrNilFactory) {
+	if err := Register(game.AbilityScorch, nil); !errors.Is(err, ErrNilFactory) {
 		t.Fatalf("expected ErrNilFactory, got %v", err)
 	}
 }
 
 func TestRegisterInvalidAbility(t *testing.T) {
 	resetRegistry(t)
-	if err := Register(shared.AbilityNone, func() AbilityHandler { return HandlerFuncs{} }); !errors.Is(err, ErrInvalidAbility) {
+	if err := Register(game.AbilityNone, func() AbilityHandler { return HandlerFuncs{} }); !errors.Is(err, ErrInvalidAbility) {
 		t.Fatalf("expected ErrInvalidAbility, got %v", err)
 	}
 }
 
 func TestNewMissingAbility(t *testing.T) {
 	resetRegistry(t)
-	if _, err := New(shared.AbilityScorch); !errors.Is(err, ErrUnknownAbility) {
+	if _, err := New(game.AbilityScorch); !errors.Is(err, ErrUnknownAbility) {
 		t.Fatalf("expected ErrUnknownAbility, got %v", err)
 	}
 }
 
 func TestNewNilHandler(t *testing.T) {
 	resetRegistry(t)
-	if err := Register(shared.AbilityScorch, func() AbilityHandler { return nil }); err != nil {
+	if err := Register(game.AbilityScorch, func() AbilityHandler { return nil }); err != nil {
 		t.Fatalf("register ability: %v", err)
 	}
-	if _, err := New(shared.AbilityScorch); !errors.Is(err, ErrNilHandler) {
+	if _, err := New(game.AbilityScorch); !errors.Is(err, ErrNilHandler) {
 		t.Fatalf("expected ErrNilHandler, got %v", err)
 	}
 }
 
 func TestRegisteredAbilitiesIsCopy(t *testing.T) {
 	resetRegistry(t)
-	if err := Register(shared.AbilityScorch, func() AbilityHandler { return HandlerFuncs{} }); err != nil {
+	if err := Register(game.AbilityScorch, func() AbilityHandler { return HandlerFuncs{} }); err != nil {
 		t.Fatalf("register ability: %v", err)
 	}
 
 	ids := registeredAbilities()
-	if len(ids) != 1 || ids[0] != shared.AbilityScorch {
+	if len(ids) != 1 || ids[0] != game.AbilityScorch {
 		t.Fatalf("unexpected ids: %v", ids)
 	}
 
 	// Mutating the returned slice should not affect the registry.
-	ids[0] = shared.AbilityDoOver
+	ids[0] = game.AbilityDoOver
 
 	registryMu.RLock()
-	_, exists := registry[shared.AbilityScorch]
-	_, mutated := registry[shared.AbilityDoOver]
+	_, exists := registry[game.AbilityScorch]
+	_, mutated := registry[game.AbilityDoOver]
 	registryMu.RUnlock()
 
 	if !exists {
